refactor: NewIdeaCard.Content 내부 Dropdown 컴포넌트 분리
