+++ src/pages/feed/components/NewIdeaCard/compound/Content/Content.tsx
@@ -2,58 +2,13 @@ import styled from '@emotion/styled';
 import { Badge, Divider, Flex, Spacer, Text, theme } from 'concept-be-design-system';
 import { MouseEventHandler, useState } from 'react';
 
-import { ReactComponent as SVGMore24 } from '../assets/more24.svg';
-import { ReactComponent as SVGPencil16 } from '../assets/pencil16.svg';
-import { ReactComponent as SVGX16 } from '../assets/x16.svg';
-import { useContentContext } from '../NewIdeaCardContext';
-
-const DropdownWrapper = styled.div`
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  top: 1.5rem;
-  right: 0;
-  width: max-content;
-  background-color: ${theme.color.w1};
-  box-shadow: 0px 6px 10px 0px rgba(0, 0, 0, 0.08);
-  border-radius: 8px;
-`;
-
-const DropdownItem = styled.button`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 6px;
-  padding: 10px;
-  cursor: pointer;
-
-  &:hover {
-    background-color: ${theme.color.l2};
-  }
-`;
-const Dropdown = () => {
-  return (
-    <DropdownWrapper>
-      <DropdownItem>
-        <Text font="suit12r" color="b6">
-          수정하기
-        </Text>
-        {/* TODO: 드롭다운에 쓰이는 연필, 삭제버튼 아이콘 디자인시스템에 추가 */}
-        <SVGPencil16 />
-      </DropdownItem>
-      <Divider color="l3" />
-      <DropdownItem>
-        <Text font="suit12r" color="b6">
-          삭제하기
-        </Text>
-        <SVGX16 />
-      </DropdownItem>
-    </DropdownWrapper>
-  );
-};
+import ContentEditDropdown from './ContentEditDropdown';
+import { ReactComponent as SVGMore24 } from '../../assets/more24.svg';
+import { useContentContext } from '../../NewIdeaCardContext';
 
 const Content = () => {
-  const { branches, title, introduce, teamRecruitments } = useContentContext();
+  const { canEdit, branches, title, introduce, teamRecruitments } = useContentContext();
+
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
 
   const isTeamRecruitmentsExist = teamRecruitments.length > 0;
@@ -84,11 +39,14 @@ const Content = () => {
 
           <Text font="suit16sb">{title}</Text>
         </Flex>
-        <Flex position="relative">
-          {/* TODO: SVGMore24 디자인시스템에 추가 */}
-          <SVGMore24 onClick={toggleDropdown} />
-          {isDropdownOpen && <Dropdown />}
-        </Flex>
+
+        {canEdit && (
+          <Flex position="relative">
+            {/* TODO: SVGMore24 디자인시스템에 추가 */}
+            <SVGMore24 onClick={toggleDropdown} />
+            {isDropdownOpen && <ContentEditDropdown />}
+          </Flex>
+        )}
       </Flex>
       <Spacer size={10} />
